requirements.txt
langchain
langchain_openai
langchain-community
streamlit
python-dotenv
google-search-results
unstructured
tiktoken
python-magic-bin
